requirements.txt
PyQt6
numpy
pillow
